+++ src/main/java/ladder/domain/result/Results.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.domain.result;
 
 import java.util.Collections;
 import java.util.List;
@@ -7,7 +7,7 @@ import java.util.stream.Stream;
 
 public class Results {
 
-    public static List<Result> results;
+    public final List<Result> results;
 
     public Results(String... strings) {
         this.results = Stream.of(strings)
