+++ src/main/java/ladder/service/LadderGameResultService.java
@@ -1,7 +1,7 @@
 package ladder.service;
 
-import ladder.domain.ResultMap;
-import ladder.domain.Results;
+import ladder.domain.result.ResultMap;
+import ladder.domain.result.Results;
 import ladder.domain.ladder.ladderline.LadderLines;
 import ladder.domain.person.Person;
