+++ src/test/java/ladder/domain/result/ResultsTest.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.domain.result;
 
 import org.junit.jupiter.api.Test;
